+++ nuxt.config.ts
@@ -1,46 +1,9 @@
 import { resolve } from 'path'
 import { config } from 'dotenv-safe'
-import { type LocaleObject } from '@nuxtjs/i18n/dist/runtime/composables'
 
 // Need to explicitly import this otherwise vite.config yells at us
 import { defineNuxtConfig } from 'nuxt/config'
-
-import { Deutsch } from './i18n/de/index'
-import { English } from './i18n/en/index'
-import { Español } from './i18n/es/index'
-import { Français } from './i18n/fr/index'
-import { 日本語 } from './i18n/ja/index'
-
-export const localeMessages = {
-  de: Deutsch,
-  en: English,
-  es: Español,
-  fr: Français,
-  ja: 日本語,
-}
-
-export const supportedLocales = [
-  {
-    code: 'de',
-    name: 'Deutsch',
-  },
-  {
-    code: 'en',
-    name: 'English',
-  },
-  {
-    code: 'es',
-    name: 'Español',
-  },
-  {
-    code: 'fr',
-    name: 'Français',
-  },
-  {
-    code: 'ja',
-    name: '日本語',
-  },
-] satisfies LocaleObject[]
+import { supportedLocales, localeMessages } from './configUtils'
 
 // https://v3.nuxtjs.org/api/configuration/nuxt.config
